#31 refactor: 유저관련 타입 정리 및 배럴 익스포트 적용
