+++ src/shared/auth/types/auth.types.ts
@@ -1,0 +1,12 @@
+import type { SessionUser } from './user.types';
+import type { Profile } from './profile.types';
+
+export type AuthState = {
+  supabaseUser: SessionUser | null;
+  profile: Profile | null;
+  hydrated: boolean;
+  setSupabaseUser: (user: SessionUser | null) => void;
+  setProfile: (profile: Profile | null) => void;
+  markHydrated: () => void;
+  reset: () => void;
+};
