+++ src/shared/auth/types/index.ts
@@ -1,0 +1,4 @@
+export * from './role.types';
+export * from './profile.types';
+export * from './user.types';
+export * from './auth.types';
