+++ src/shared/auth/types/profile.types.ts
@@ -1,4 +1,4 @@
-import type { Role } from './role';
+import type { Role } from './role.types';
 
 export type Profile = {
   uuid: string;
