+++ src/shared/auth/types/user.types.ts
